vendoredvulns/frontend/pyproject.toml
[tool.poetry]
name = "vendored-vulns-frontend"
version = "0.1.0"
description = "Frontend component for the nixpkgs vendored vulnerabilities dashboard."
authors = ["Pierre Bourdon <delroth@gmail.com>"]
license = "MIT"
readme = "../../README.md"
packages = [{include = "vendoredvulns/frontend", from = "../.."}]

[tool.poetry.scripts]
vendoredvulns-frontend-manage = "vendoredvulns.frontend.manage:main"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
django-debug-toolbar = "^4.2.0"
django-allauth = "^0.58.2"
pygithub = "^2.1.1"
cachetools = "^5.3.2"
urllib3 = "^2.1.0"
django-simple-history = "^3.4.0"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
